requirements.txt
requests==2.25.1
websocket-client
pandas
numpy